Remove own implementation of popup
